Extract and Analyze Rivers from Elevation Data.

Seamless extraction of river networks from digital elevation models data. The package allows analysis of digital elevation models that can be either externally provided or downloaded from open source repositories (thus interfacing with the elevatr package). Extraction is performed via the 'D8' flow direction algorithm of TauDEM (Terrain Analysis Using Digital Elevation Models), thus interfacing with the traudem package. Resulting river networks are compatible with functions from the OCNet package.

Author

Luca Carraro (luca.carraro@hotmail.it)